pyproject.toml
[project]
name = "desafio-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "NicolasSP90",email = "nicolassp90@hotmail.com"}
]
readme = "README.md"
requires-python = "3.13.*"
dependencies = [
    "django[bcrypt] (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
